requirements.txt
langchain
openai
huggingface_hub
python-dotenv
langchain-community
langchain-openai